Refactor: rename to e2e_test
