+++ Makefile
@@ -9,7 +9,7 @@ all: $(TARGET)
 
 check:
 	golangci-lint run
-	$(MAKE) -C adaptors_test/keys/testdata
+	$(MAKE) -C e2e_test/keys/testdata
 	go test -v -race -cover -coverprofile=coverage.out ./...
 
 $(TARGET): $(wildcard *.go)
